requirements.txt
alembic
annotated-types
anyio
bcrypt==4.3.0
certifi
cffi
charset-normalizer
click
cryptography
distlib
ecdsa
exceptiongroup
fastapi
filelock
greenlet
h11
httpcore
httpx
idna
iniconfig
Mako
MarkupSafe
numpy
opencv-python-headless
packaging
passlib==1.7.4
patchright
pillow
platformdirs
playwright
pluggy
psycopg2-binary
pyasn1
pycparser
pydantic
pydantic_core
pyee
Pygments
pyotp
pytesseract
pytest
pytest-asyncio
python-dotenv
python-jose
python-multipart
requests
rsa
six
sniffio
SQLAlchemy
starlette
tomli
typing-inspection
typing_extensions
urllib3
uvicorn
virtualenv
